+++ src/features/widgets/components/TokenBalanceWidget/TokenBalanceWidget.tsx
@@ -1,19 +1,16 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import {
   Divider,
   Flex,
   Skeleton,
   SkeletonCircle,
   Text,
-  TextProps,
   useColorModeValue,
 } from '@chakra-ui/react';
 
 import { abbreviateNumber } from 'js-abbreviation-number';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCaretUp, faCaretDown } from '@fortawesome/free-solid-svg-icons';
-import { SizeProp } from '@fortawesome/fontawesome-svg-core';
 import { WidgetWrapper } from '../WidgetWrapper/WidgetWrapper';
+import { ChangeIndicator } from '../../../shared/components/ChangeIndicator/ChangeIndicator';
 import { TokenBalanceWidgetSettings } from './TokenBalanceWidgetSettings/TokenBalanceWidgetSettings';
 
 export interface Token {
@@ -36,81 +33,6 @@ export interface TokenBalanceWidgetProps {
   historicalBalance: Balance;
   isLoading: boolean;
 }
-
-export interface ChangeIndicatorProps {
-  size: 'sm' | 'lg';
-  trend: 'upwards' | 'downwards';
-}
-
-export const ChangeIndicator: React.FC<ChangeIndicatorProps> = ({
-  size,
-  trend,
-}) => {
-  const sizing = useMemo<{
-    iconSize: SizeProp;
-    fontSize: TextProps['fontSize'];
-  }>(() => {
-    switch (size) {
-      case 'lg':
-        return {
-          iconSize: '1x',
-          fontSize: 'xs',
-        };
-
-      case 'sm':
-        return {
-          iconSize: 'xs',
-          fontSize: 'x-small',
-        };
-    }
-  }, [size]);
-  const upwardsTrendColor = useColorModeValue('green.500', 'green.500');
-  const downwardsTrendColor = useColorModeValue('red.500', 'red.500');
-  const trendColor = useMemo(() => {
-    switch (trend) {
-      case 'upwards':
-        return upwardsTrendColor;
-      case 'downwards':
-        return downwardsTrendColor;
-    }
-  }, [trend, upwardsTrendColor, downwardsTrendColor]);
-  const trendIcon = useMemo(() => {
-    switch (trend) {
-      case 'upwards':
-        return faCaretUp;
-      case 'downwards':
-        return faCaretDown;
-    }
-  }, [trend]);
-  const trendIconOffset = useMemo(() => {
-    switch (trend) {
-      case 'upwards':
-        return '0.5px';
-      case 'downwards':
-        return '-0.5px';
-    }
-  }, [trend]);
-  return (
-    <Flex
-      alignItems={'center'}
-      color={trendColor}
-      justifyContent={'center'}
-      pl={'2'}
-    >
-      <Flex alignItems={'center'} justifyContent={'center'} pr={'1'}>
-        <FontAwesomeIcon
-          icon={trendIcon}
-          size={sizing.iconSize}
-          style={{
-            position: 'relative',
-            top: trendIconOffset,
-          }}
-        />
-      </Flex>
-      <Text fontSize={sizing.fontSize}>(+36%)</Text>
-    </Flex>
-  );
-};
 
 export const TokenBalanceWidget: React.FC<TokenBalanceWidgetProps> = ({
   balance,
